backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.0
SQLAlchemy==2.0.35
alembic==1.13.3
jinja2==3.1.4
reportlab==4.2.2

torch==2.3.1
torchvision==0.18.1
torchaudio==2.3.1
pytorch-lightning==2.4.0
captum==0.7.0
Pillow==10.4.0
scikit-learn==1.5.2
numpy==1.26.4
pandas==2.2.3

pytest==8.3.3
httpx==0.27.2

black==24.8.0
flake8==7.1.1
isort==5.13.2






